// src/app/shared/models/stock-quote.ts
export interface IStockQuote {
  company: string;
  symbol: string;
  c?: number; //Current price
  d?: number; //Change
  dp?: number; //Percent change
  h?: number; //High price of the day
  l?: number; //Low price of the day
  o?: number; //Open price of the day
  pc?: number; //Previous close price
  t?: number; //Timestamp
}